+++ src/components/createEvent/detail/detail.tsx
@@ -1,5 +1,5 @@
 import { Dispatch, SetStateAction, useState } from "react";
-import styles from "../../../styles/event/create.module.css";
+import InputDetail from "./inputDetail";
 
 export default function EventDetail({
   detail,
@@ -8,37 +8,9 @@ export default function EventDetail({
   detail: string;
   setDetail: Dispatch<SetStateAction<string>>;
 }) {
-  const [error, setError] = useState("");
-
-  const handleEventDetailChange = (e) => {
-    setDetail(e.target.value);
-    // 一文字でも入力されたらエラー削除
-    if (error) {
-      setError("");
-    }
-  };
-
-  const handleEventNameBlur = () => {
-    if (detail.trim() === "") {
-      setError("※詳細内容は入力必須項目です");
-    } else {
-      setError("");
-    }
-  };
   return (
     <>
-      <textarea
-        className={`${styles.detail} ${error ? styles.errorInput : ""} `}
-        maxLength={250}
-        value={detail}
-        onChange={handleEventDetailChange}
-        onBlur={handleEventNameBlur}
-      />
-      {error && (
-        <div>
-          <span className={styles.span}>{error}</span>
-        </div>
-      )}
+      <InputDetail detail={detail} setDetail={setDetail} />
     </>
   );
 }
