+++ src/components/createEvent/detail/inputDetail.tsx
@@ -1,0 +1,33 @@
+// JSX表示のみコンポーネント
+
+import { Dispatch, SetStateAction, useState } from "react";
+import styles from "../../../styles/event/create.module.css";
+import HandleDetailBlur from "./handleDetailBlur";
+import HandleChange from "./handleChange";
+
+export default function InputDetail({
+  detail,
+  setDetail,
+}: {
+  detail: string;
+  setDetail: Dispatch<SetStateAction<string>>;
+}) {
+  const [error, setError] = useState("");
+
+  return (
+    <>
+      <textarea
+        className={`${styles.detail} ${error ? styles.errorInput : ""} `}
+        maxLength={250}
+        value={detail}
+        onChange={HandleChange({ setDetail, error, setError })}
+      />
+      {error && (
+        <div>
+          <span className={styles.span}>{error}</span>
+        </div>
+      )}
+      <HandleDetailBlur detail={detail} setError={setError} />
+    </>
+  );
+}
